requirements.txt
aiounittest==1.4.2
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.2
asyncpg==0.29.0
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==41.0.7
dill==0.3.7
distlib==0.3.8
fastapi==0.108.0
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
mirakuru==2.5.2
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
port-for==0.7.2
pre-commit==3.6.0
psutil==5.9.7
psycopg==3.1.17
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pylint==3.0.3
pytest==7.4.4
pytest-postgresql==5.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
setuptools==69.0.3
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
tomlkit==0.12.3
trio==0.24.0
trustme==1.1.0
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
